--- src/main/java/cn/lanqiao/HospitalInpatient/Interceptors/RedirectInterceptor.java
@@ -1,29 +1,0 @@
-package cn.lanqiao.HospitalInpatient.Interceptors;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.web.servlet.HandlerInterceptor;
-import org.springframework.web.servlet.ModelAndView;
-import java.io.IOException;
-
-public class RedirectInterceptor implements HandlerInterceptor {
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        String requestUri = request.getRequestURI();
-        if ("/".equals(requestUri) || "/index.html".equals(requestUri)) {
-            response.sendRedirect("/login.html");
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
-        // 这里暂不处理，可根据需要添加逻辑
-    }
-
-    @Override
-    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
-        // 这里暂不处理，可根据需要添加逻辑
-    }
-}
