--- src/main/java/cn/lanqiao/HospitalInpatient/Interceptors/WebMvcConfig.java
@@ -1,15 +1,0 @@
-package cn.lanqiao.HospitalInpatient.Interceptors;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new MyInterceptor())
-                .addPathPatterns("/"); // 拦截所有请求
-    }
-}
