+++ src/main/java/cn/lanqiao/HospitalInpatient/config/MyInterceptor.java
@@ -1,0 +1,29 @@
+package cn.lanqiao.HospitalInpatient.config;
+
+import org.springframework.web.servlet.HandlerInterceptor;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.web.servlet.ModelAndView;
+
+public class MyInterceptor implements HandlerInterceptor {
+
+    // 请求处理之前执行
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+        System.out.println("请求路径: " + request.getRequestURI());
+        // 返回 true 表示继续执行后续的处理（即进入 Controller），返回 false 则表示请求被拦截，后续不再执行
+        return true;
+    }
+
+    // 请求处理之后，视图渲染之前执行
+    @Override
+    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
+        System.out.println("请求处理完毕，返回视图之前");
+    }
+
+    // 请求处理完毕，视图渲染之后执行
+    @Override
+    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
+        System.out.println("请求处理完毕，响应返回之后");
+    }
+}
