+++ src/main/java/cn/lanqiao/HospitalInpatient/config/WebMvcConfiguration.java
@@ -6,6 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.List;
@@ -24,5 +25,13 @@ public class WebMvcConfiguration implements WebMvcConfigurer {
         converter.setObjectMapper(new JacksonObjectMapper());
 //        将自己的消息转换器加入容器
         converters.add(0,converter);
+    }
+
+    // 拦截器
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new MyInterceptor())
+                .addPathPatterns("/"). // 拦截所有请求
+                excludePathPatterns("/login"); // 排除登录请求
     }
 }
