+++ src/main/java/cn/lanqiao/HospitalInpatient/service/AdminUserService.java
@@ -5,5 +5,4 @@ import cn.lanqiao.HospitalInpatient.model.pojo.Admin;
 import com.baomidou.mybatisplus.extension.service.IService;
 
 public interface AdminUserService extends IService<Admin> {
-//    Admin login(Admin admin);
 }
