+++ src/main/resources/application.yml
@@ -1,10 +1,15 @@
 server:
  port: 8080
 spring:
+  mvc:
+    sataic - path - patterns: /static/**
+  web:
+    resources:
+      static-locations: classpath:/static/
   datasource:
     username: root
     password: 20010720Lmx
-    url: jdbc:mysql://47.109.206.82:3306/hospital?allowPublicKeyRetrieval=true&useSSL=false&serverTimezone=UTC&characterEncoding=utf8&characterSetResults=utf8
+    url: jdbc:mysql://47.109.206.2:3306/hospital?allowPublicKeyRetrieval=true&useSSL=false&serverTimezone=UTC&characterEncoding=utf8&characterSetResults=utf8
 #  thymeleaf:
 #    cache: false
 #    check-template-location: 'true'
