+++ src/main/resources/static/left.html
@@ -300,6 +300,13 @@
 </table>
 <script src="js/left.js"></script> <!-- 假设把JavaScript代码放在script.js中 -->
 <script>
+  if (window.self!== window.top) {
+    // 页面在iframe中，正常显示
+  } else {
+    // 页面被单独访问，重定向到错误页面
+    window.location.href = 'error.html';
+  }
+
   // 删除登录信息
   function deleteCookie(name) {
     document.cookie = name + '=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;';
@@ -321,7 +328,7 @@
       whichEl.style.display = "none";
     }
   }
-</script> <!-- 假设把JavaScript代码放在script.js中 -->
+</script>
 </body>
 
 </html>
